+++ examples/Playground.hs
@@ -53,7 +53,9 @@ testT = Test
     eitherT1 = Toml.match (Toml._Integer >>> Toml._Left)  "either.Left"
            <|> Toml.match (Toml._String  >>> Toml._Right) "either.Right"
 
-    -- same key for sum type
+    -- same key for sum type;
+    -- doesn't work if you have something like `Either String String`,
+    -- you should distinguish these cases by different keys like in `eitherT1` example
     eitherT2 :: BiToml (Either String Double)
     eitherT2 = ( Toml.match (Toml._String >>> Toml._Left)
              <!> Toml.match (Toml._Double >>> Toml._Right)