+++ src/Toml.hs
@@ -8,16 +8,16 @@ import qualified Toml
 
 module Toml
     ( module Toml.Bi
+    , module Toml.BiMap
     , module Toml.Parser
     , module Toml.PrefixTree
     , module Toml.Printer
-    , module Toml.Prism
     , module Toml.Type
     ) where
 
 import Toml.Bi
+import Toml.BiMap
 import Toml.Parser
 import Toml.PrefixTree
 import Toml.Printer
-import Toml.Prism
 import Toml.Type