+++ CHANGELOG.md
@@ -4,6 +4,15 @@ Change log
 tomland uses [PVP Versioning][1].
 The change log is available [on GitHub][2].
 
+0.4.0
+=====
+
+* [#54](https://github.com/kowainik/tomland/issues/54):
+  Add support for sum types.
+  Rename `Prism` to `BiMap`.
+  Rename `bijectionMaker` to `match`.
+  Add `string` codec.
+
 0.3.1
 =====
 * [#19](https://github.com/kowainik/tomland/issues/19):